Better creation API for RoboAttributeSet
