+++ robolectric-shadows/shadows-core/src/main/java/org/robolectric/fakes/RoboAttributeSet.java
@@ -1,7 +1,11 @@
-package org.robolectric.shadows;
+package org.robolectric.fakes;
 
+import android.content.Context;
 import android.util.AttributeSet;
 import android.util.TypedValue;
+
+import com.google.android.collect.Lists;
+
 import org.robolectric.RuntimeEnvironment;
 import org.robolectric.res.AttrData;
 import org.robolectric.res.Attribute;
@@ -10,8 +14,11 @@ import org.robolectric.res.ResType;
 import org.robolectric.res.ResourceIndex;
 import org.robolectric.res.ResourceLoader;
 import org.robolectric.res.TypedResource;
+import org.robolectric.shadows.Converter;
 
 import java.util.List;
+
+import static org.robolectric.Shadows.shadowOf;
 
 /**
  * Robolectric implementation of {@link android.util.AttributeSet}.
@@ -25,10 +32,12 @@ public class RoboAttributeSet implements AttributeSet {
     this.resourceLoader = resourceLoader;
   }
 
+  @Deprecated
   public RoboAttributeSet put(String fullyQualifiedName, String value, String valuePackage) {
     return put(new Attribute(fullyQualifiedName, value, valuePackage));
   }
 
+  @Deprecated
   public RoboAttributeSet put(Attribute attribute) {
     attributes.add(attribute);
     return this;
@@ -193,6 +202,14 @@ public class RoboAttributeSet implements AttributeSet {
     return i != null ? i : 0;
   }
 
+  /**
+   * Creates a {@link RoboAttributeSet} for the given {@link Attribute}(s)
+   */
+  public static RoboAttributeSet create(Context context, Attribute... attrs) {
+    List<Attribute> attributesList = Lists.newArrayList(attrs);
+    return new RoboAttributeSet(attributesList,
+        shadowOf(context.getAssets()).getResourceLoader());
+  }
 
   private ResName getAttrResName(String namespace, String attrName) {
     String packageName = Attribute.extractPackageName(namespace);
